경로 조회에 대한 코드 리팩토링
