+++ src/main/java/nextstep/subway/applicaion/PathService.java
@@ -2,6 +2,7 @@ package nextstep.subway.applicaion;
 
 import nextstep.subway.applicaion.dto.PathResponse;
 import nextstep.subway.domain.*;
+import nextstep.subway.domain.subwaymap.DistanceSubwayMap;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,8 +20,9 @@ public class PathService {
     public PathResponse findPath(Long source, Long target, PathType type) {
         Station upStation = stationService.findById(source);
         Station downStation = stationService.findById(target);
+
         List<Line> lines = lineService.findLines();
-        SubwayMap subwayMap = new SubwayMap(lines);
+        DistanceSubwayMap subwayMap = new DistanceSubwayMap(lines);
         Path path = subwayMap.findPath(upStation, downStation);
 
         return PathResponse.of(path);
