+++ src/main/java/nextstep/subway/domain/subwaymap/SubwayMap.java
@@ -1,0 +1,10 @@
+package nextstep.subway.domain.subwaymap;
+
+import nextstep.subway.domain.Path;
+import nextstep.subway.domain.Station;
+
+public interface SubwayMap {
+
+    Path findPath(Station source, Station target);
+
+}
